requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
pydantic==2.5.3
alembic==1.13.1
python-dotenv==1.0.0
gunicorn==21.2.0
